# API/Models/AuthResponse.cs
namespace WebApi.Models;

public class AuthResponse
{
    public string Token { get; set; } = string.Empty;
    public DateTime ExpiresAt { get; set; }
    public User User { get; set; } = new User();
} 