# API/Models/CreateUserDto.cs
using System.ComponentModel.DataAnnotations;

namespace WebApi.Models;

public class CreateUserDto
{
    [Required]
    [StringLength(50, MinimumLength = 3)]
    [RegularExpression(@"^[a-zA-Z0-9_]+$", ErrorMessage = "Username can only contain letters, numbers, and underscores")]
    public string Username { get; set; } = string.Empty;

    [Required]
    [EmailAddress]
    [StringLength(255)]
    public string Email { get; set; } = string.Empty;

    [Required]
    [StringLength(100, MinimumLength = 6)]
    public string Password { get; set; } = string.Empty;

    [Required]
    [StringLength(100)]
    public string FirstName { get; set; } = string.Empty;

    [Required]
    [StringLength(100)]
    public string LastName { get; set; } = string.Empty;

    [StringLength(20)]
    public string Role { get; set; } = "User";

    [Url]
    public string? ProfileImageUrl { get; set; }

    [StringLength(1000)]
    public string? Bio { get; set; }

    public DateTime? DateOfBirth { get; set; }
} 