# news_swagger.json
{
    "openapi": "3.0.1",
    "info": {
    "title": "News API",
    "description": "A comprehensive News API with JWT authentication for posting and managing news articles with MongoDB",
    "contact": {
    "name": "News API Support",
    "email": "support@newsapi.com"
    },
    "version": "v1"
    },
    "paths": {
    "/api/Auth/register": {
    "post": {
    "tags": [
    "Auth"
    ],
    "summary": "Register a new user and return JWT token",
    "requestBody": {
    "description": "User registration data",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/CreateUserDto"
    }
    },
    "text/json": {
    "schema": {
    "$ref": "#/components/schemas/CreateUserDto"
    }
    },
    "application/*+json": {
    "schema": {
    "$ref": "#/components/schemas/CreateUserDto"
    }
    }
    }
    },
    "responses": {
    "201": {
    "description": "User registered successfully with JWT token",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AuthResponse"
    }
    }
    }
    },
    "400": {
    "description": "Invalid input data or username/email already exists",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "500": {
    "description": "Internal server error"
    }
    }
    }
    },
    "/api/Auth/login": {
    "post": {
    "tags": [
    "Auth"
    ],
    "summary": "Authenticate user and return JWT token",
    "requestBody": {
    "description": "Login credentials",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/LoginDto"
    }
    },
    "text/json": {
    "schema": {
    "$ref": "#/components/schemas/LoginDto"
    }
    },
    "application/*+json": {
    "schema": {
    "$ref": "#/components/schemas/LoginDto"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Authentication successful with JWT token",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AuthResponse"
    }
    }
    }
    },
    "400": {
    "description": "Invalid input data",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "401": {
    "description": "Invalid credentials",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "500": {
    "description": "Internal server error"
    }
    }
    }
    },
    "/api/Auth/profile": {
    "get": {
    "tags": [
    "Auth"
    ],
    "summary": "Get current user profile (requires authentication)",
    "responses": {
    "200": {
    "description": "Returns current user profile",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/User"
    }
    }
    }
    },
    "401": {
    "description": "Authentication required",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "404": {
    "description": "User not found",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "500": {
    "description": "Internal server error"
    }
    }
    },
    "put": {
    "tags": [
    "Auth"
    ],
    "summary": "Update current user profile (requires authentication)",
    "requestBody": {
    "description": "User update data",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UpdateUserDto"
    }
    },
    "text/json": {
    "schema": {
    "$ref": "#/components/schemas/UpdateUserDto"
    }
    },
    "application/*+json": {
    "schema": {
    "$ref": "#/components/schemas/UpdateUserDto"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Profile updated successfully",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/User"
    }
    }
    }
    },
    "400": {
    "description": "Invalid input data",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "401": {
    "description": "Authentication required",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "404": {
    "description": "User not found",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "500": {
    "description": "Internal server error"
    }
    }
    }
    },
    "/api/Auth/validate-token": {
    "post": {
    "tags": [
    "Auth"
    ],
    "summary": "Validate JWT token",
    "requestBody": {
    "description": "Token validation request",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/RefreshTokenRequest"
    }
    },
    "text/json": {
    "schema": {
    "$ref": "#/components/schemas/RefreshTokenRequest"
    }
    },
    "application/*+json": {
    "schema": {
    "$ref": "#/components/schemas/RefreshTokenRequest"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Token is valid",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    },
    "400": {
    "description": "Invalid or expired token",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "500": {
    "description": "Internal server error"
    }
    }
    }
    },
    "/api/Auth/refresh": {
    "post": {
    "tags": [
    "Auth"
    ],
    "summary": "Refresh JWT token (requires authentication)",
    "responses": {
    "200": {
    "description": "Token refreshed successfully",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AuthResponse"
    }
    }
    }
    },
    "401": {
    "description": "Authentication required",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "404": {
    "description": "User not found",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ProblemDetails"
    }
    }
    }
    },
    "500": {
    "description": "Internal server error"
    }
    }
    }
    },
...
    },
    "components": {
    "schemas": {
    "AuthResponse": {
    "type": "object",
    "properties": {
    "token": {
    "type": "string",
    "nullable": true
    },
    "expiresAt": {
    "type": "string",
    "format": "date-time"
    },
    "user": {
    "$ref": "#/components/schemas/User"
    }
    },
    "additionalProperties": false
    },
...
    "CreateUserDto": {
    "required": [
    "email",
    "firstName",
    "lastName",
    "password",
    "username"
    ],
    "type": "object",
    "properties": {
    "username": {
    "maxLength": 50,
    "minLength": 3,
    "pattern": "^[a-zA-Z0-9_]+$",
    "type": "string"
    },
    "email": {
    "maxLength": 255,
    "minLength": 0,
    "type": "string",
    "format": "email"
    },
    "password": {
    "maxLength": 100,
    "minLength": 6,
    "type": "string"
    },
    "firstName": {
    "maxLength": 100,
    "minLength": 0,
    "type": "string"
    },
    "lastName": {
    "maxLength": 100,
    "minLength": 0,
    "type": "string"
    },
    "role": {
    "maxLength": 20,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "profileImageUrl": {
    "type": "string",
    "format": "uri",
    "nullable": true
    },
    "bio": {
    "maxLength": 1000,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "dateOfBirth": {
    "type": "string",
    "format": "date-time",
    "nullable": true
    }
    },
    "additionalProperties": false
    },
    "LoginDto": {
    "required": [
    "password",
    "usernameOrEmail"
    ],
    "type": "object",
    "properties": {
    "usernameOrEmail": {
    "maxLength": 255,
    "minLength": 0,
    "type": "string"
    },
    "password": {
    "maxLength": 100,
    "minLength": 0,
    "type": "string"
    }
    },
    "additionalProperties": false
    },
...
    "ProblemDetails": {
    "type": "object",
    "properties": {
    "type": {
    "type": "string",
    "nullable": true
    },
    "title": {
    "type": "string",
    "nullable": true
    },
    "status": {
    "type": "integer",
    "format": "int32",
    "nullable": true
    },
    "detail": {
    "type": "string",
    "nullable": true
    },
    "instance": {
    "type": "string",
    "nullable": true
    }
    },
    "additionalProperties": {}
    },
    "RefreshTokenRequest": {
    "required": [
    "token"
    ],
    "type": "object",
    "properties": {
    "token": {
    "minLength": 1,
    "type": "string"
    }
    },
    "additionalProperties": false
    },
...
    "UpdateUserDto": {
    "type": "object",
    "properties": {
    "username": {
    "maxLength": 50,
    "minLength": 3,
    "pattern": "^[a-zA-Z0-9_]+$",
    "type": "string",
    "nullable": true
    },
    "email": {
    "maxLength": 255,
    "minLength": 0,
    "type": "string",
    "format": "email",
    "nullable": true
    },
    "firstName": {
    "maxLength": 100,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "lastName": {
    "maxLength": 100,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "role": {
    "maxLength": 20,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "isActive": {
    "type": "boolean",
    "nullable": true
    },
    "profileImageUrl": {
    "type": "string",
    "format": "uri",
    "nullable": true
    },
    "bio": {
    "maxLength": 1000,
    "minLength": 0,
    "type": "string",
    "nullable": true
    },
    "dateOfBirth": {
    "type": "string",
    "format": "date-time",
    "nullable": true
    }
    },
    "additionalProperties": false
    },
    "User": {
    "type": "object",
    "properties": {
    "id": {
    "type": "string",
    "nullable": true
    },
    "username": {
    "type": "string",
    "nullable": true
    },
    "email": {
    "type": "string",
    "nullable": true
    },
    "passwordHash": {
    "type": "string",
    "nullable": true
    },
    "firstName": {
    "type": "string",
    "nullable": true
    },
    "lastName": {
    "type": "string",
    "nullable": true
    },
    "role": {
    "type": "string",
    "nullable": true
    },
    "isActive": {
    "type": "boolean"
    },
    "profileImageUrl": {
    "type": "string",
    "nullable": true
    },
    "bio": {
    "type": "string",
    "nullable": true
    },
    "dateOfBirth": {
    "type": "string",
    "format": "date-time",
    "nullable": true
    },
    "createdAt": {
    "type": "string",
    "format": "date-time"
    },
    "updatedAt": {
    "type": "string",
    "format": "date-time"
    },
    "lastLoginAt": {
    "type": "string",
    "format": "date-time",
    "nullable": true
    }
    },
    "additionalProperties": false
    }
    },
    "securitySchemes": {
    "Bearer": {
    "type": "apiKey",
    "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.",
    "name": "Authorization",
    "in": "header"
    }
    }
    },
    "security": [
    {
    "Bearer": []
    }
    ]
    }